third_party/WebKit/LayoutTests/media/track/vtt-cue-exceptions-expected.txt

Verify that VTTCue exceptions are properly messaged to developers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cue.vertical = 'Invalid!'; threw exception SyntaxError: Failed to set the 'vertical' property on 'VTTCue': The value provided ('Invalid!') is invalid. Only 'rl', 'lr', and the empty string are accepted..
PASS cue.line = -1; threw exception IndexSizeError: Failed to set the 'line' property on 'VTTCue': The snap-to-lines flag is not set, and the value provided (-1) is not between 0 and 100..
PASS cue.line = 101; threw exception IndexSizeError: Failed to set the 'line' property on 'VTTCue': The snap-to-lines flag is not set, and the value provided (101) is not between 0 and 100..
PASS cue.position = -1; threw exception IndexSizeError: Failed to set the 'position' property on 'VTTCue': The value provided (-1) is not between 0 and 100..
PASS cue.position = 101; threw exception IndexSizeError: Failed to set the 'position' property on 'VTTCue': The value provided (101) is not between 0 and 100..
PASS cue.size = -1; threw exception IndexSizeError: Failed to set the 'size' property on 'VTTCue': The value provided (-1) is not between 0 and 100..
PASS cue.size = 101; threw exception IndexSizeError: Failed to set the 'size' property on 'VTTCue': The value provided (101) is not between 0 and 100..
PASS cue.align = 'Invalid!'; threw exception SyntaxError: Failed to set the 'align' property on 'VTTCue': The value provided ('Invalid!') is invalid. Only 'start', 'middle', 'end', 'left', and 'right' are accepted..
PASS successfullyParsed is true

TEST COMPLETE
